updated test location
