+++ particula/util/tests/convert_test.py
@@ -242,44 +242,3 @@ def test_effective_refractive_index():
             volume_zero=10,
             volume_one=5,
         ) == (1.4572585227821824+0.3214931829339477j)
-
-
-def test_datetime64_from_epoch_array():
-    """Test the datetime64_from_epoch_array function."""
-    # Test with a simple example array
-    epoch_array = np.array([0, 1, 2, 3, 4])
-    expected_result = np.array(['1970-01-01T00:00:00', '1970-01-01T00:00:01',
-                                '1970-01-01T00:00:02', '1970-01-01T00:00:03',
-                                '1970-01-01T00:00:04'], dtype='datetime64[s]')
-    assert np.array_equal(
-            convert.datetime64_from_epoch_array(epoch_array),
-            expected_result
-        )
-
-    # Test with a non-zero delta
-    delta = 3600  # 1 hour in seconds
-    expected_result = np.array(['1970-01-01T01:00:00', '1970-01-01T01:00:01',
-                                '1970-01-01T01:00:02', '1970-01-01T01:00:03',
-                                '1970-01-01T01:00:04'], dtype='datetime64[s]')
-    assert np.array_equal(
-            convert.datetime64_from_epoch_array(epoch_array, delta),
-            expected_result
-        )
-
-    # Test with an empty array
-    empty_array = np.array([])
-    try:
-        convert.datetime64_from_epoch_array(empty_array)
-        assert False, \
-            "Function should raise an AssertionError for empty array."
-    except AssertionError:
-        pass
-
-    # Test with a non-integer array
-    float_array = np.array([0.0, 1.0, 2.0])
-    try:
-        convert.datetime64_from_epoch_array(float_array)
-        assert False, \
-            "Function should raise an AssertionError for non-integer array."
-    except AssertionError:
-        pass
